+++ CompletedExamples/ComputingDiscreteLogarithms.py
@@ -24,8 +24,8 @@ Date of Initial Creation: November 1, 2021
 
 """
 
-from MiscFunctions import FactoringFunctions as FactorFuncs
 from ShorFactoringAlgorithm import ShorFactoringAlgorithm as ShorFA
+from src.MiscFunctions import FactoringFunctions as FactorFuncs
 
 __author__ = "Alex Lim"
 __credits__ = "Alex Lim"
