+++ TestFile.py
@@ -11,8 +11,8 @@ Date of Initial Creation: October 19, 2021
 
 import numpy as np
 
-from QuantumCircuitMatrix import QuantumCircuitMatrix as qcm
-from ShorFactoringAlgorithm import ShorFactoringAlgorithm
+from src.QuantumCircuitMatrix import QuantumCircuitMatrix as qcm
+from CompletedExamples.ShorFactoringAlgorithm import ShorFactoringAlgorithm
 
 __author__      = "Alex Lim"
 __credits__     = "Alex Lim"
