+++ WIP/bogosort/QuantumBogoSort.py
@@ -14,10 +14,9 @@ Date of Initial Creation: October 21, 2021
 import numpy as np
 from numpy import dot as i, kron as d, ndarray as k
 
-from bogosort import bogosort
-from MiscFunctions import MiscFunctions as idk
-# from numpy import identity as eeee
-from QuantumCircuitMatrix import QuantumCircuitMatrix as e
+from WIP.bogosort.bogosort import bogosort
+from src.MiscFunctions import MiscFunctions as idk
+from src.QuantumCircuitMatrix import QuantumCircuitMatrix as e
 
 __author__ = "Alex Lim"
 __credits__ = "Alex Lim"
